+++ ogc-example/src/main/scala/geotrellis/server/ogc/conf/package.scala
@@ -27,9 +27,11 @@ import com.azavea.maml.ast.codec.tree._
 import com.typesafe.config._
 import geotrellis.raster.io.geotiff.{Auto, AutoHigherResolution, Base, OverviewStrategy}
 import io.circe._
+import io.circe.syntax._
 import io.circe.parser._
+import cats.syntax.either._
 import pureconfig._
-import pureconfig.error.CannotConvert
+import pureconfig.error.{CannotConvert, ExceptionThrown}
 import pureconfig.generic.FieldCoproductHint
 import pureconfig.generic.auto._
 
@@ -184,4 +186,10 @@ package object conf {
       case _                              => OverviewStrategy.DEFAULT
     }
   }
+
+  implicit val ogcTimeFormatReader: ConfigReader[OgcTimeFormat] =
+    ConfigReader[String].emap { _.asJson.as[OgcTimeFormat].leftMap(e => ExceptionThrown(e.fillInStackTrace())) }
+
+  implicit val ogcTimeDefaultReader: ConfigReader[OgcTimeDefault] =
+    ConfigReader[String].emap { _.asJson.as[OgcTimeDefault].leftMap(e => ExceptionThrown(e.fillInStackTrace())) }
 }