+++ stac-example/src/main/scala/geotrellis/server/ogc/stac/package.scala
@@ -22,10 +22,11 @@ import geotrellis.store.query._
 
 import geotrellis.raster.{EmptyName, RasterSource, SourceName, StringName}
 import geotrellis.raster.geotiff.GeoTiffPath
-import com.azavea.stac4s.StacItemAsset
+import com.azavea.stac4s.{StacExtent, StacItemAsset}
 import com.azavea.stac4s.api.client.{SearchFilters, StacClient, Query => SQuery}
 import io.circe.syntax._
 import cats.{Applicative, Foldable, FunctorFilter}
+import cats.data.NonEmptyList
 import cats.syntax.either._
 import cats.syntax.foldable._
 import cats.syntax.functor._
@@ -33,7 +34,19 @@ import cats.syntax.functorFilter._
 import cats.syntax.applicative._
 import eu.timepit.refined.types.string.NonEmptyString
 
+import java.time.{ZoneOffset, ZonedDateTime}
+
 package object stac {
+  implicit class StacExtentionOps(val self: StacExtent) extends AnyVal {
+
+    /** [[StacExtent]]s with no temporal component are valid. */
+    def ogcTime: Option[OgcTime] = self.temporal.interval.headOption.map(_.value.flatten.map(_.atZone(ZoneOffset.UTC))).map {
+      case fst :: Nil        => OgcTimeInterval(fst)
+      case fst :: snd :: Nil => OgcTimeInterval(fst, snd)
+      case _                 => OgcTimeEmpty
+    }
+  }
+
   implicit class StacSummaryOps(val self: StacSummary) extends AnyVal {
     def sourceName: SourceName =
       self match {
@@ -72,7 +85,7 @@ package object stac {
     /** A helper function that filters raster sources in case the STAC Layer is temporal and it is not taken into account in the query.
       *
       * By default STAC API returns all temporal items even though the time is not specified.
-      * If defaultTime configuration is set to true and the query is not temporal and not universal
+      * If ignoreTime configuration is set to false and the query is not temporal and not universal
       * (meaning that it is bounded by temporal or spatial extent),
       * we can select the first time position of the temporal layer in this case.
       *
@@ -80,15 +93,14 @@ package object stac {
       * All non temporal items would be included into the result.
       * Otherwise, only items that match the first time position would be returned.
       */
-    def timeSlice(query: Query, defaultTime: Boolean, datetimeField: Option[String]): G[T] =
-      if (defaultTime && query.nonTemporal && query.nonUniversal) {
+    def timeSlice(query: Query, timeDefault: OgcTimeDefault, ignoreTime: Boolean, datetimeField: Option[String]): G[T] =
+      if (!ignoreTime & query.nonTemporal && query.nonUniversal) {
         self.foldMap(_.time(datetimeField)) match {
-          case OgcTimePositions(list)       =>
-            val start = list.head
-            self.filter(source => OgcTime.strictTimeMatch(source.time(datetimeField), start))
-          case OgcTimeInterval(start, _, _) =>
-            self.filter(source => OgcTime.strictTimeMatch(source.time(datetimeField), start))
-          case OgcTimeEmpty                 => self
+          case OgcTimePositions(list)         =>
+            self.filter(source => OgcTime.strictTimeMatch(source.time(datetimeField), timeDefault.selectTime(list)))
+          case OgcTimeInterval(start, end, _) =>
+            self.filter(source => OgcTime.strictTimeMatch(source.time(datetimeField), timeDefault.selectTime(NonEmptyList.of(start, end))))
+          case OgcTimeEmpty                   => self
         }
       } else self