+++ CHANGELOG.md
@@ -20,6 +20,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - StacCollectionRasterSource implementation [#340](https://github.com/geotrellis/geotrellis-server/issues/340)
 - WCS Rendering formats support [#195](https://github.com/geotrellis/geotrellis-server/issues/195)
 - GeoTrellis Server STAC Package [#350](https://github.com/geotrellis/geotrellis-server/issues/350)
+- STAC Collection temporal extents support [#347](https://github.com/geotrellis/geotrellis-server/issues/347)
+- OGC Services configurable time representation [#324](https://github.com/geotrellis/geotrellis-server/issues/324)
+- Default time period configurable on the WMS layer with a time dimension [#325](https://github.com/geotrellis/geotrellis-server/issues/325)
 
 ### Changed
 - Update GT Server STAC4S dependency [#319](https://github.com/geotrellis/geotrellis-server/issues/319)